refactor: move zip stream fetcher to infra package
