+++ application/src/test/java/run/halo/app/core/extension/theme/ThemeEndpointTest.java
@@ -40,6 +40,7 @@ import run.halo.app.extension.ReactiveExtensionClient;
 import run.halo.app.infra.SystemConfigurableEnvironmentFetcher;
 import run.halo.app.infra.SystemSetting;
 import run.halo.app.infra.ThemeRootGetter;
+import run.halo.app.infra.ZipStreamFetcher;
 import run.halo.app.theme.TemplateEngineManager;
 
 /**
